moved ExhaustPluginInterface to src\types\interface.ts, added ExhaustPluginStorage type.
