+++ src/types/exhaust-plugin-interface.ts
@@ -1,6 +1,6 @@
 import {Context} from './manifest';
 
-export interface ExhaustPluginInterface {
+export interface ExhaustPluginInterface_deleteMe {
   /**
    * Execute exhaust based on `context` and `tree`, produce output to a file in `outputPath`.
    */
