+++ src/types/initialize.ts
@@ -1,5 +1,9 @@
-import {PluginInterface} from './interface';
+import {ExhaustPluginInterface, PluginInterface} from './interface';
 
 export type PluginsStorage = {
   [key: string]: PluginInterface;
+};
+
+export type ExhaustPluginStorage = {
+  [key: string]: ExhaustPluginInterface;
 };
