+++ src/types/interface.ts
@@ -1,3 +1,5 @@
+import {Context} from './manifest';
+
 export type PluginParams = Record<string, any>;
 
 export type PluginInterface = {
@@ -10,3 +12,10 @@ export type PluginInterface = {
   };
   [key: string]: any;
 };
+
+export interface ExhaustPluginInterface {
+  /**
+   * Execute exhaust based on `context` and `tree`, produce output to a file in `outputPath`.
+   */
+  execute(tree: any, context: Context, outputPath?: string): void;
+}
